Split sdk-macros into separate modules
